+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v0.3.0 - unreleased
 
+- Change `on_missing` option in the `priority_filter` from "error" to "raise".
+  ([#79](https://github.com/mathause/filefinder/pull/79))
 - Drop support for python 3.7 and 3.8 ([#80](https://github.com/mathause/filefinder/pull/80))
 - Allow passing scalar numbers to `find_paths` and `find_files` ([#58](https://github.com/mathause/filefinder/issues/58)).
 - Show duplicates for non-unique queries